Use a redis mutex to ensure channel messages are published sequentially
MessageBus publishing happens outside of the lua scripts, and is therefore not atomic. For example, if a user were to leave and then join a channel in quick succession, it is essential that the messagebus messages are published in order of `leave, join`.

This commit teaches the present/leave/auto_leave lua scripts to obtain a mutex if required, and adds a method to release the mutex after the messagebus logic has finished in Ruby. The count/user_id methods do not need to obtain the mutex, but they do check that nothing else is holding the mutex. This is important to ensure the last_message_id is not off-by-one.